requirements.txt
flask
requests
watchdog
langchain
sentence-transformers
pinecone-client
python-dotenv
ollama
gradio